feat(S2): Отработаны всевозможные ситуации, на них вроде работает
