+++ mihalchenko.vladislav/S2/stack.hpp
@@ -6,7 +6,6 @@
 // #include <iostream>
 // #include <string>
 // #include <memory>
-#include <set>
 #include <stack>
 
 namespace mihalchenko
